变强每日数据结构体Playerid使用UInt64,用Int是错误的

+++ Core/GData/GDataDBExecHelper.h
@@ -471,7 +471,7 @@ struct DBReportId
 };
 struct DBDaysRank
 {
-    int player_id;
+    UInt64 player_id;
     std::string athletics;
     std::string level;
     std::string money;
@@ -1024,7 +1024,7 @@ SPECIALEND()
 SPECIALBEGIN(GData::DBDaysRank)
 SPECIALDEF(6)
 	(
-	int,player_id,
+	UInt64,player_id,
     std::string, athletics,
     std::string, level,
     std::string, money,
